requirements.txt
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dill==0.3.5.1
Flask==2.1.3
fonttools==4.34.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.5.12
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyaml==21.10.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.10
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.1
